requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.0.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
httpie==1.0.2
idna==2.7
pkg-resources==0.0.0
Pygments==2.3.0
PyJWT==1.6.4
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
votes==1.0.1
